pyproject.toml
[project]
name = "agentic-resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.47.2",
    "openai>=1.109.1",
    "pypdf>=6.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
